requirements.txt
aiogram==3.7.0
groq>=1.0.0
yt-dlp==2024.12.23
python-dotenv==1.0.1
aiofiles==23.2.1
